+++ TESTING.cs
@@ -1,3 +1,4 @@
+using Dawn;
 using HarmonyLib;
 using SnowyLib;
 using UnityEngine;
@@ -27,6 +28,7 @@ namespace Rats
             try
             {
                 if (!Utils.testing) { return; }
+
             }
             catch
             {